Corrected flake8, currentLocation removed from _popuMenuLocation, function mapset_exists moved from utils.py to guiutils.py.
